Create csc.util package
