+++ src/main/java/redis/clients/jedis/util/JedisURIHelper.java
@@ -5,6 +5,8 @@ import redis.clients.jedis.csc.ClientSideCache;
 import redis.clients.jedis.HostAndPort;
 import redis.clients.jedis.Protocol;
 import redis.clients.jedis.RedisProtocol;
+import redis.clients.jedis.csc.util.GuavaCSC;
+import redis.clients.jedis.csc.util.CaffeineCSC;
 
 public final class JedisURIHelper {
